pyproject.toml
[project]
name = "venv"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "accelerate>=1.13.0",
    "aocd>=0.1",
    "astunparse>=1.6.3",
    "beautifulsoup4>=4.14.3",
    "dotenv>=0.9.9",
    "gensim>=4.4.0",
    "google-genai>=1.73.1",
    "huggingface>=0.0.1",
    "jsonlines>=4.0.0",
    "numpy>=2.4.4",
    "openai>=2.32.0",
    "pandas>=3.0.2",
    "pyarrow>=24.0.0",
    "pytest>=9.0.3",
    "python-dotenv>=1.2.1",
    "pyyml>=0.0.2",
    "requests>=2.32.5",
    "tenacity>=9.1.4",
    "tiktoken>=0.12.0",
    "tqdm>=4.67.3",
    "transformers>=5.6.2",
]
